Arreglado tema cookies a falta de probarlo en production

+++ src/app/cookies/cookie-consent.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { SsrCookieService } from 'ngx-cookie-service-ssr';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root',
@@ -7,26 +8,35 @@ import { SsrCookieService } from 'ngx-cookie-service-ssr';
 export class CookieConsentService {
   private consentCookieName = 'cookie_consent';
 
-  constructor(private ssrCookieService: SsrCookieService) {}
+  constructor(
+    private ssrCookieService: SsrCookieService,
+    @Inject(PLATFORM_ID) private platformId: any
+  ) {}
 
   hasConsent(): boolean {
+    if (!isPlatformBrowser(this.platformId)) return true; // Evita verificar cookies en SSR
     return this.ssrCookieService.check(this.consentCookieName);
   }
 
   setConsent(value: 'accepted' | 'rejected'): void {
-    this.ssrCookieService.set(this.consentCookieName, value, {
-      expires: 365,
-      path: '/',
-      secure: true,
-      sameSite: 'Lax' // Usa 'None' si necesitas cross-site cookies
-    });
+    if (isPlatformBrowser(this.platformId)) {
+      this.ssrCookieService.set(this.consentCookieName, value, {
+        expires: 365,
+        path: '/',
+        secure: true,
+        sameSite: 'Lax',
+      });
+    }
   }
 
   getConsent(): string {
+    if (!isPlatformBrowser(this.platformId)) return ''; // Evita errores en SSR
     return this.ssrCookieService.get(this.consentCookieName);
   }
 
   revokeConsent(): void {
-    this.ssrCookieService.delete(this.consentCookieName, '/');
+    if (isPlatformBrowser(this.platformId)) {
+      this.ssrCookieService.delete(this.consentCookieName, '/');
+    }
   }
 }
